Rename files to match CapCasing
