+++ src/VideoStream.cpp
@@ -1,4 +1,4 @@
-#include "videoStream.hpp"
+#include "VideoStream.hpp"
 
 
 VideoStream::VideoStream()
